+++ src/components/Root/FFetch.test.js
@@ -3,6 +3,7 @@
 /* eslint-disable no-unused-vars */
 
 import ms from 'ms';
+import '../../../test/jest/__mock__';
 
 import { getTokenExpiry } from '../../loginServices';
 import { FFetch } from './FFetch';
@@ -24,8 +25,21 @@ const log = jest.fn();
 
 const mockFetch = jest.fn();
 
+const mockBroadcastChannel = {
+  postMessage: jest.fn(),
+  onmessage: null,
+  close: jest.fn(),
+  addEventListener: jest.fn((event, callback) => {
+    if (event === 'message') {
+      mockBroadcastChannel.onmessage = callback;
+    }
+  }),
+  removeEventListener: jest.fn(),
+};
+
 describe('FFetch class', () => {
   beforeEach(() => {
+    global.BroadcastChannel = jest.fn(() => mockBroadcastChannel);
     global.fetch = mockFetch;
     getTokenExpiry.mockResolvedValue({
       atExpires: Date.now() + (10 * 60 * 1000),
@@ -40,11 +54,13 @@ describe('FFetch class', () => {
   describe('Calling a non-FOLIO API', () => {
     it('calls native fetch once', async () => {
       mockFetch.mockResolvedValueOnce('non-okapi-success');
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -57,11 +73,13 @@ describe('FFetch class', () => {
   describe('Calling a FOLIO API fetch', () => {
     it('calls native fetch once', async () => {
       mockFetch.mockResolvedValueOnce('okapi-success');
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
       const response = await global.fetch('okapiUrl/whatever', { testOption: 'test' });
@@ -90,7 +108,7 @@ describe('FFetch class', () => {
         store: {
           dispatch: jest.fn(),
         },
-        okapi:{
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
         }
@@ -112,11 +130,13 @@ describe('FFetch class', () => {
   describe('logging out', () => {
     it('calls native fetch once to log out', async () => {
       mockFetch.mockResolvedValueOnce('logged out');
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -130,11 +150,13 @@ describe('FFetch class', () => {
     it('fetch failure is silently trapped', async () => {
       mockFetch.mockRejectedValueOnce('logged out FAIL');
 
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -153,11 +175,13 @@ describe('FFetch class', () => {
   describe('logging out', () => {
     it('Calling an okapi fetch with valid token...', async () => {
       mockFetch.mockResolvedValueOnce('okapi success');
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -202,7 +226,7 @@ describe('FFetch class', () => {
         rtrConfig: {
           fixedLengthSessionWarningTTL: '1m',
         },
-        okapi:{
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
         }
@@ -262,7 +286,7 @@ describe('FFetch class', () => {
         rtrConfig: {
           fixedLengthSessionWarningTTL: '1m',
         },
-        okapi:{
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
         }
@@ -304,7 +328,7 @@ describe('FFetch class', () => {
         rtrConfig: {
           fixedLengthSessionWarningTTL: '1m',
         },
-        okapi:{
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
         }
@@ -344,7 +368,7 @@ describe('FFetch class', () => {
         store: {
           dispatch: jest.fn(),
         },
-        okapi:{
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
         }
@@ -391,7 +415,7 @@ describe('FFetch class', () => {
         rtrConfig: {
           fixedLengthSessionWarningTTL: '1m',
         },
-        okapi:{
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
         }
@@ -423,11 +447,13 @@ describe('FFetch class', () => {
     it('returns the error', async () => {
       mockFetch.mockResolvedValue('success')
         .mockResolvedValueOnce('failure');
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -449,11 +475,13 @@ describe('FFetch class', () => {
             },
           }
         ));
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -477,11 +505,13 @@ describe('FFetch class', () => {
           }
         ))
         .mockRejectedValueOnce(new Error('token error message'));
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -516,11 +546,13 @@ describe('FFetch class', () => {
             }
           }
         ));
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -550,11 +582,13 @@ describe('FFetch class', () => {
             }
           }
         ));
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -584,11 +618,13 @@ describe('FFetch class', () => {
             }
           }
         ));
-      const testFfetch = new FFetch({ logger: { log },
-        okapi:{
+      const testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
           url: 'okapiUrl',
           tenant: 'okapiTenant'
-        } });
+        }
+      });
       testFfetch.replaceFetch();
       testFfetch.replaceXMLHttpRequest();
 
@@ -598,6 +634,37 @@ describe('FFetch class', () => {
         expect(e instanceof UnexpectedResourceError).toBeTrue;
         expect(mockFetch.mock.calls).toHaveLength(0);
       }
+    });
+  });
+
+  describe('active window messaging', () => {
+    let testFfetch;
+    beforeEach(() => {
+      testFfetch = new FFetch({
+        logger: { log },
+        okapi: {
+          url: 'okapiUrl',
+          tenant: 'okapiTenant'
+        }
+      });
+      testFfetch.replaceFetch();
+      testFfetch.replaceXMLHttpRequest();
+    });
+
+    it('sends a message when setActiveWindow is called', async () => {
+      const windowId = window.stripesRTRWindowId;
+      testFfetch.documentFocusHandler();
+      expect(mockBroadcastChannel.postMessage).toHaveBeenCalledWith({
+        type: '@folio/stripes/core::activeWindowMessage',
+        activeWindow: windowId,
+      });
+    });
+
+    it('handles messages from other windows', async () => {
+      const windowId = 'test-window-id';
+      mockBroadcastChannel.onmessage({ data: { type: '@folio/stripes/core::activeWindowMessage', activeWindow: windowId } });
+
+      expect(sessionStorage.getItem('@folio/stripes/core::activeWindowId')).toEqual(windowId);
     });
   });
 });